Added a describe command, which presents information about the addon for the user.
Also, made some changes to list command, now it prints the names of the addon folders,
rather than the project names. This makes it possible to use this information easily
in other commands, such as describe.

Functions for extracting Curse Project Name and Version fields from .toc files have
been added as well. These should make it possible to actually check if there is an
update or not first before downloading updates.